backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.1
SQLAlchemy==2.0.29
alembic==1.13.1
psycopg[binary]==3.1.18
python-dotenv==1.0.1
pydantic-settings==2.2.1
openai==1.54.3
google-generativeai==0.8.3
httpx>=0.27.0
